+++ Tests/PSScriptAnalyzer.Tests.ps1
@@ -20,4 +20,4 @@ foreach ($PowerShellFile in $PowerShellFiles) {
             }
         }
     }
-}# foreach
+} # foreach